Rename main application to: example
The old name (helloc) suggested that the driver code
was in the helloc.o file instead of main.o. Also, make
implicit rules worked only because helloc.o was a dependency.

Print PROGRAM and VERSION with the driver code instead of
helloc_world().  The message is the same so we preserve
backward compatibility (helloc_world()'s message was never
explicitly specified).